+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 Release Notes
 =============
+## To Be Decided
+* Web App: Support for adding site extensions with "use_extension"
+
 ## 1.3.0-beta1
 * Container Instance: Support for command line arguments.
 * Key Vault: Support for setting tags on key vault secrets.